Fixed bug #432: Downtime scheduled as "Nagios Process" and not the Users name
Previously the comment associated with a scheduled downtime would use a 
hard-coded value of "(Nagios Process)" as the user that scheduled the 
downtime. Now it will use the author from the downtime itself in the comment
associated with the downtime.

If no downtime author is specified (or more accurately, the author field in the
downtime structure is NULL), the system will default to "(Nagios Process)"
as the author of the downtime.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/branches/nagios-3-4-x@2667 5f96b256-904b-4d8d-8c98-d829582c6739