Fixed bug #445: Adding triggered downtime for child hosts causes a SIGSEGV on restart/reload
This was caused by triggered downtimes being deleted the when the
triggering downtime was restarted. It was deleted because it was
still marked as in effect. It is now marked as not in effect in the
register_downtime() function.

A related issue, also resolved, is that after a restart, the
triggered downtime was dropped. The same issue also caused the CGI
not to list the triggered downtime. This was due to the ordering
of the downtimes in the retention.dat and status.dat files.

Previously the triggered downtime always appeared before its
triggering downtime in those files. When the downtimes were read
from those files, either on a core restart or by the CGIs, the
triggered downtime would be discarded because the triggering
downtime did not yet exist.

The most common case for this is when a downtime is created and
the option is selected to create triggered downtimes on all child
objects. A change was made in the way downtimes are sorted so that
triggered downtimes with the same start times as untriggered 
downtimes always appear later in the list.  This change in the 
sort order does NOT resolve the case where a manually created, 
triggered downtime is created with a start time earlier than the 
triggering downtime.

This would need to be resolved by comparing the triggered_by value
with the downtime ID regardless of the start time. However, this
should be a relatively rare case and only caused by intentional
scheduling by a human. This change was not implemented because it
would cause the downtime list to be out of time order and the
implications of this were not well understood.



git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/branches/nagios-3-4-x@2752 5f96b256-904b-4d8d-8c98-d829582c6739