Fixes for CGI localization
Previously we were checking in the CGIs to see if any of the languages
passed in the Accept-Language HTTP header were supported on the local
system, but we were not stopping at the first one that was supported.
Now we are.

Also, we previously set all aspects of localization (LC_ALL in the
setlocale call) based on the Accept-Language. This caused problems
reading status.dat and object.cache files for some localizations.
Specifically, if a localization uses a comma for the decimal point
rather than a period, the CGI might report that it could not read
the file. This was resolved by setting only the LC_CTYPE aspects
of locale based on the Accept-Language.

A longer term solution might be to enable localization in core
(which it is not), but that would require configuration files to
use the localizations as well. This is a bigger change that needs
additional thought.